requirements.txt
Pillow==3.4